requirements.txt
Django>=3.0.6,<3.1.0
djangorestframework>=3.11.0,<3.12.0
psycopg2>=2.8.5,<2.9.0
Pillow>=7.1.2,<7.2.0
flake8>=3.7.9,<3.8.0
drf-yasg>=1.17.1,<1.18.0
gunicorn>=20.0.4,<20.1.0
runscript>=0.2.9,<0.3
django-extensions>=2.2.9,<2.3
Faker>=4.1.0,<4.2
django-cors-headers>=3.3.0,<3.4.0